[CHORE] : balance 타입 수정

+++ app/src/main/java/com/sopt/kakaobank/data/dto/response/ResponseAccountDto.kt
@@ -8,7 +8,7 @@ data class ResponseAccountDto(
     @SerialName("accountName")
     val accountName: String,
     @SerialName("balance")
-    val balance: Int,
+    val balance: Long,
     @SerialName("accountId")
     val accountId: Int
 )
